# docs/swagger.yaml
basePath: /
definitions:
...
  serializer.CompleteCalculationRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  serializer.InsulinCalculationCartInfoRespJSON:
    properties:
      insulin_calculation_id:
        type: integer
      patient_count:
        type: integer
    type: object
  serializer.InsulinCalculationJSON:
    properties:
      calculated_at:
        type: string
      comment:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      creator_id:
        type: string
      insulin_calculation_id:
        type: integer
      moderator_id:
        type: string
      status:
        type: string
    required:
    - creator_id
    - status
    type: object
  serializer.InsulinCalculationRespJSON:
    properties:
      calculated_at:
        type: string
      comment:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      creator_username:
        type: string
      insulin_calculation_id:
        type: integer
      moderator_username:
        type: string
      status:
        type: string
    type: object
  serializer.InsulinCalculationWithPatientsRespJSON:
    properties:
      insulin_calculation:
        $ref: '#/definitions/serializer.InsulinCalculationRespJSON'
      patient_count:
        type: integer
      patients:
        items:
          $ref: '#/definitions/serializer.PatientInCalculationRespJSON'
        type: array
    type: object
  serializer.PatientInCalculationRespJSON:
    properties:
      bread_units:
        type: number
      calculated_insulin:
        type: number
      current_glucose:
        type: number
      name:
        type: string
      sensitivity:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@insulin.ru
    name: API Support
  description: API для расчета болюсного инсулина
  license:
    name: AS IS (NO WARRANTY)
  title: Insulin Calculation API
  version: "1.0"
paths:
  /api/insulin-calculations:
    get:
      consumes:
      - application/json
      description: Возвращает список расчетов с фильтрацией по статусу и дате
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Начальная дата фильтрации
        in: query
        name: start_date
        type: string
      - description: Конечная дата фильтрации
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список расчетов
          schema:
            items:
              $ref: '#/definitions/serializer.InsulinCalculationRespJSON'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список расчетов инсулина
      tags:
      - insulin-calculations
  /api/insulin-calculations/{insulin_calculation_id}:
    delete:
      consumes:
      - application/json
      description: Удаляет расчет инсулина (только черновики)
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Расчет успешно удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID расчета
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить расчет
      tags:
      - insulin-calculations
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию о расчете включая список пациентов
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные расчета с пациентами
          schema:
            $ref: '#/definitions/serializer.InsulinCalculationWithPatientsRespJSON'
        "400":
          description: Неверный ID расчета
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить расчет с пациентами
      tags:
      - insulin-calculations
    put:
      consumes:
      - application/json
      description: Обновляет данные расчета инсулина
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/serializer.InsulinCalculationJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Расчет успешно обновлен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные расчета
      tags:
      - insulin-calculations
  /api/insulin-calculations/{insulin_calculation_id}/complete:
    put:
      consumes:
      - application/json
      description: Завершает или отклоняет расчет модератором
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      - description: Статус завершения
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/serializer.CompleteCalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Расчет успешно завершен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить или отклонить расчет
      tags:
      - insulin-calculations
  /api/insulin-calculations/{insulin_calculation_id}/form:
    put:
      consumes:
      - application/json
      description: Переводит расчет из статуса черновика в сформированный
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Расчет успешно сформирован
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка формирования
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать расчет
      tags:
      - insulin-calculations
...
  /api/insulin-calculations/info:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о текущем черновике расчета (ID и количество
        пациентов)
      produces:
      - application/json
      responses:
        "200":
          description: Информация о корзине
          schema:
            $ref: '#/definitions/serializer.InsulinCalculationCartInfoRespJSON'
        "404":
          description: Черновик не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить информацию о корзине расчета
      tags:
      - insulin-calculations
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"